Preview Version 0.1.0 (#1)
* Finished the design for the globe app bar.

Now it shows the time, has a expand screen button, and adapts to mobile screens.
Made some extra pinia stores for the time and screen size functionality.

* Added a Resume Page.

This page takes little space and saves someone from Making a GET request to an S3 bucket, ultimately saving money.
Added a resume icon to the Globe App Bar as well.

* Adjusted the colors of the search bar to a blue theme.

Also completely removed the old globe bar formerly at the bottom.
Made some files to set up the new menus for the globe.

* Made the search menu.

This comes with initial styles for any app menu.
Logic hasn't been made yet.

* Made a more mobile-friendly home screen.

The app bar is on the left instead of the bottom.

* Adjusted the resume sidebar.

This makes up for some visual complications.
Fixed an issue with the geocoder.

* Made the logic for the search menu.

Now it works properly with the rest of the code.

* Made updates to the menu.

Now, the menu has a new design with blue and pink colors.
Made an incomplete menu as a placeholder for menus that aren't complete.
Removed a bunch of old code. I am going to start over.

* Set Up the Time and Settings Menu.

Made the Globe App bar work a bit better.
Made some improvements to the Mount functions.

* Started and Finished making the Settings Menu.

There are many actions in the settings menu now.
Also added new functions and Icons to this menu.

* uninstalled the brand icons.

Also updated the version number.

* Setup GitHub Actions

Also removed the Time Menu for the moment.