Version Update 0.4.0 (#4)
* Updated and added MANY dependencies.

This will help auto import components and utilities, locally host font family files, and more.

* Updated imports with new plugins now.

* Added FA brand icons + removed all external assets.

I wasn't using any of the assets. Also, using the brand library provided by Font awesomeis much easier.

* Added the Map Settings Menu.

This simply will be used to set the map and some traits on the map.
Also changed the layout of certain buttons on the app bar.

* Multiple features added. See Description.

Added options to the map settings menu to change the map layers being used. It takes heavy inspiration from Worlds iVue.
Updated the Options Menu and the App Bar with links to the repository for this website.

* Added an option to show the Google 3D Tileset to the Globe.