Version Update 0.5.0 (#5)
* Added Tauri to my app.

This will make it easier to deploy my globe platform as a desktop app.

* uninstalled tauri.

Its reliance on Rust can make it harder to use for the future. Imma switch to electron.

* started to get electron to work with the app.

* Removed electron now - It is MASSIVE

* Added tauri and got it to build for a windows OS

Turns out Tauri is better than Electron for my purposes.

* Made some final edits.

Removed Plenty of useless assets.
Changed Version Number to 0.5.0.
Added a button to the app bar to download the app on the app bar.