Version Update 0.5.1 (#6)
* Made many updates. See Description

Updated dependencies and version number. Also uninstalled old electron dev dependency.
Added a new about menu. no content in it yet, but it has been properly integrated into the app.
Made an alias for a search menu. any menu can now have an alias route.
Made a custom scrollbar for the menus.

* Added Natural Earth native layers.

* Corrected the size of the layer buttons on mobile devices.

Also added the build:both command for building the web frontend and the tauri app frontend.

* Finished making the about menu.

The Map settings menu whenever new map layers are added.

* Made an interval that detects when the app has an internet connection.

It changes the map layers based on that.

* Last Changes for Version Update 0.5.1. See Description.

Made an .ico version of one of my icons so the online interval works better.
Changed the Version number of my app in src-tauri.
Rebuilt the app for windows.