Version Update 0.5.2 (#7)
* SIGNIFICANTLY updated dependencies.

They were months old at this point and needed to be updated.
Also, I track whether the mode for development is with Tauri or not in a better way.

* Updated the OG vite config file to accomodate for the Tauri mode.

Edited Package.json and added settings.json as well.

* Made some redirect routes.

They work like the ones on my main website.
I also updated some configuration logic so that there is type support for the font awesome icons and better logic for the fullscreen store and FA Icons.

* Built the app with Tauri using the correct version number.